Move dbscan to utils
